boards: openhwgroup: add CV64A6 on GenesysII board
Adds support for the CVA6 CPU on a GenesysII FPGA board
(https://github.com/openhwgroup/cva6).
The SoC currently contains the CVA6 CPU in 64-bit configuration with the
SV39 MMU, interrupt controllers (CLINT and PLIC), UART, a SPI for
booting from SD, a boot ROM, and I2C controller for on-board audio, a
GPIO and the lowRISC ethernet subsystem.

Signed-off-by: Eric Ackermann <eric.ackermann@cispa.de>